moved course and user controller to controller package
