--- src/main/java/org/launchcode/sprinklespre/controllers/UserController.java
@@ -1,4 +1,0 @@
-package org.launchcode.sprinklespre.controllers;
-
-public class UserController {
-}
